refact: refactoring dtos for course
